Added english tutorial for Grafana monitoring (#323)
* Added english tutorial for Grafana monitoring

I have written a tutorial on how to install Grafana on Ubuntu 20.04. It uses InfluxDB and Telegraf, there is also a corresponding dashboard.

* Updated Dashboard import Picture

Due to a privacy problem

* Removed tmux lines from example

for a better formatting

* removed misleading line

I had a leftover from a configuration with several databases in the tutorial. This is not necessary with this variant with one database server and a single database.

* Added tutorail pictures for permanent availability

-

* Reviewed Tutorial

- Added header
- removed sudo lines
- reuploaded pictures and changes the links
- Added explanation line for twice added repos
- removed double repo from Grafana
- removed port from URL on additional Telegraf setup, since these requests get redirected by the proxy

* Added  **Prerequisites**

* changed ending headline to ## Conclusion

* Added MIT license block & corrected typo

* Reworked Lets-Encrypt Part & corrected image links

I reworked the tutorials Certificate Part. Now i refer to the tutorial of https://github.com/dschoeffm. Which uses the Hetzner-DNS API for create and renewal. I think its a better way, to always have a valid certificate.

* Created Tutorial for Ngninx-Proxymanager

Started with the Tutorial

* Added Basic Auth Part - corrected some typos

* Update 01.en.md

* Added additional Information for Subdomains

Also formatted some Entries

* Update 01.en.md

* Update 01.en.md

* Update 01.en.md

* Deleted additonal tutorial from this pull request

I have accidentally when creating an additional tutorial, this with on this pull request gemerged. Should be adjusted accordingly again. I'm still learning how to use Github.

* Formatting and spelling updates

Co-authored-by: leojakeson <47002532+leojakeson@users.noreply.github.com>