ubuntu-24-visualisierung-with-QEMU (#1311)
* ubuntu-24-visualisierung-with-QEMU

* Update date in Ubuntu 24 QEMU tutorial

* Update date in Ubuntu 24 QEMU tutorial

* Formatting updates

* Add proper introduction

* Split tutorials

* Revise German text in QEMU bridging tutorial

Updated the German translation for clarity and accuracy in the tutorial on bridging with QEMU.

* Update conclusion in Ubuntu virtualization tutorial

Clarified the conclusion about the bridge setup and its benefits.

* Clarify outcome of routed setup in documentation

Updated the outcome section to clarify the network setup and its benefits.

* Update 01.de.md

* Refine content and add conclusion for clarity

Updated the text to improve clarity and added a conclusion summarizing the configuration benefits.

---------

Co-authored-by: svenja.michal <svenja.michal@hetzner-cloud.de>